+++ R/R/l_plot_density.R
@@ -12,9 +12,12 @@
 #' @param ylabel the graphical parameter ylabel labelling the y axis of the plot. If \code{NULL} (the default), an ylabel is created
 #' based on the combination of data name and "density".
 #' @param title providing a title for the plot. If \code{NULL} (the default), the title will be the call which produced the result.
-#' @param linewidth line width, default value is 2.
-#' @param color line colour.  If \code{NULL} (the default) colour will be taken from \code{l_getOption("foreground")}.
-#' @param size points size, default value is 0.5
+#' @param linewidth line width.
+#' Default is given by \code{\link{l_getOption}("linewidth")}.
+#' @param color line colour.
+#' Default is given by \code{\link{l_getOption}("color")}.
+#' @param size points size.
+#' Default is given by \code{\link{l_getOption}("size")}.
 #' @param ... named arguments being states passed to l_plot()
 #'
 #' @examples
@@ -29,11 +32,13 @@
 #' @export
 
 l_plot.density <- function(x, y = NULL, xlabel = NULL,  ylabel = NULL, title = NULL,
-                           linewidth = 2, color = NULL, size = 0.5,
+                           linewidth = l_getOption("linewidth"),
+                           color = l_getOption("color"),
+                           size = l_getOption("size"),
                            ...){
     if (!is.null(y)) warning("value of y argument is ignored")
     density <- x
-    if (is.null(color)) color <- l_getOption("foreground")
+    if (is.null(color)) color <- l_getOption("color")
     if(is.null(ylabel)){ylabel <- paste(density$data.name, "density")}
     if(is.null(xlabel)){xlabel <- paste("N =", density$n, " ", "Bandwidth =", round(density$bw, 3))}
     if(is.null(title)){title <- paste("density of", density$data.name)}